implementata frequenza come numero a 16 bit, aumentato il ventaglio di frequenze disponibili dando all'utente la possibilità di scegliere il periodo in decimi di ms. Possibilità di scegliere tra due BaudRate differenti

+++ AVR/avr_common/my_adc.h
@@ -6,5 +6,5 @@
 void adc_init(void);
 void timer_init(void);
 void adc_sel(uint8_t adc_number);
-void freq_set(uint8_t frequency);
+void freq_set(uint16_t frequency);
 void adc_conv(uint8_t var);
